requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.2

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0

# SAR/GIS Processing
rasterio>=1.2.0
GDAL>=3.3.0
sentinelsat>=1.1.0
snappy>=8.0.0

# API
fastapi>=0.68.0
uvicorn>=0.15.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.1.0

# Jupyter
jupyter>=1.0.0
notebook>=6.4.0

# Utils
pyyaml>=5.4.0
tqdm>=4.62.0 